fix basic UI
